Estandarizacion de codigo
