+++ src/app/[type]/page.tsx
@@ -1,0 +1,41 @@
+'use client';
+import { useEffect, useState } from "react";
+import { useParams } from "next/navigation";
+import { getDataFilter} from "@lib/getData";
+import { Data} from "@app-types/types";
+import Acordion from "@components/ui/Acordion";
+import TotalForCategory from "@components/TotalForCategory";
+
+type Params = {
+  type: string;
+}
+export default function Page() {
+const params = useParams() as Params;
+const {type} = params;
+
+const [data, setData] = useState< Data[] | null >(null);
+
+useEffect(()=>{
+  const fetchData = async () => {
+    getDataFilter(type).then(setData);
+  }
+  fetchData();    
+},[type]);
+
+const colors = {
+    savings: "text-savings",
+    expenses: "text-expenses",
+    income: "text-income",
+  };
+const color = colors[type] || "text-primary";
+
+return (
+  <section className=" flex flex-col items-center">
+    <h2 className={`${color} font-extrabold text-2xl `}>
+        {type.toUpperCase()} $<TotalForCategory type={type} />
+    </h2>
+    <Acordion data={data} type={type} />
+  </section>
+);
+}
+
