+++ src/app/cards/addCard.tsx
@@ -1,0 +1,10 @@
+import Form from "@components/cards/Form";
+
+export default function FormCard() {
+  return (
+    <section>
+      <h2 className="text-primary font-bold text-xl text-center mb-5">Agregar Tarjeta</h2>
+      <Form/>
+    </section> 
+  );
+}
