+++ src/app/cards/page.tsx
@@ -1,0 +1,7 @@
+export default function CardsPage() {
+  return (
+    <>
+    <h2 className="text-red-700 text-2xl font-extrabold m-3">Tarjetas</h2>
+    </>
+  )
+}
