--- src/app/formCard/page.tsx
@@ -1,10 +1,0 @@
-import CardForm from "@components/CardForm";
-
-export default function Formpage() {
-  return (
-    <section>
-      <h2 className="text-primary font-bold text-xl text-center mb-5">Agregar Tarjeta</h2>
-      <CardForm/>
-    </section> 
-  );
-}
