+++ src/app/layout.tsx
@@ -1,19 +1,22 @@
 import type {Metadata} from "next";
-import Header from "@components/Header";
-import "./globals.css";
+import "@app-styles/globals.css";
+import Header from "@components/layouts/Header";
+import Footer from "@components/layouts/Footer";
 
-export const metadata: Metadata ={
-  title: "Cuentas App",
-  description: "App create by next app",
-}
+export const metadata: Metadata = {
+  title: 'Home Financial App',
+  description: 'Administration of personal finances',
+};
 
-export default function RootLayout({children,}: Readonly<{children: React.ReactNode;}>) {
+export default function RootLayout({children,}: {children: React.ReactNode;}) {
   return (
     <html lang="en" data-theme="light">
-      <body className="w-full h-screen relative">
+      <body className="min-h-screen">
         <Header/>
-        <main className="my-3 pb-[80px]">{children}</main>
-   
+        <main>
+          {children}
+        </main>
+        <Footer/>
       </body>
     </html>
   );
