+++ src/app/page.tsx
@@ -1,11 +1,17 @@
-import Home from "@components/Home";
-import AddItemsHome from "@components/AddItemsHome";
+import AddCategory from "@components/home/AddCategory";
+import Graph from '@components/home/Graph';
+import InfoList from '@components/home/InfoList';
 
-export default function HomePage() {
+export default function App() {
+  const currentMonth = new Date().toLocaleString('en', { month: 'long' });
   return (
-    <div>
-      <AddItemsHome/>
-      <Home/>
-    </div>
+    <>
+      <AddCategory/>
+      <div className="flex flex-col items-center justify-between w-full h-full">
+        <h2 className="text-red-700 text-2xl font-extrabold m-3">{currentMonth}</h2>
+        <Graph/>
+        <InfoList/>
+      </div>
+    </>
   );
 }
