+++ src/components/layouts/Footer.tsx
@@ -1,0 +1,8 @@
+export default function Footer() {
+  return (
+    <footer className="absolute bottom-0 left-0 pl-3 w-full h-10 shadow-xl shadow-gray-300/60 flex justify-between items-center bg-header text-white text-sm">  
+      <h2 className="w-full">Financial Home</h2>
+      <p className="text-[.8em] w-full">© 2025 Financial Home. All rights reserved.</p>
+    </footer>
+  );
+}
